refactor: split modifiers.ts into focused modules
Split the 485-line modifiers.ts into focused modules to reduce merge
conflicts when multiple agents work in parallel:

- modifiers/queries.ts: getModifiersOfType, getModifiersForPlayer, etc.
- modifiers/combat.ts: getEffectiveEnemyArmor/Attack, isAbilityNullified
- modifiers/terrain.ts: getEffectiveTerrainCost
- modifiers/cardValues.ts: getEffectiveSidewaysValue, isRuleActive
- modifiers/lifecycle.ts: addModifier, removeModifier, expireModifiers
- modifiers/index.ts: barrel export for backward compatibility
