+++ packages/core/src/engine/__tests__/blackAsAnyColor.test.ts
@@ -7,7 +7,7 @@
 
 import { describe, it, expect } from "vitest";
 import { validateManaAvailable } from "../validators/mana/index.js";
-import { addModifier } from "../modifiers.js";
+import { addModifier } from "../modifiers/index.js";
 import { createTestGameState, createTestPlayer } from "./testHelpers.js";
 import type { ManaSource, SourceDie } from "../../state/GameState.js";
 import {
