+++ packages/core/src/engine/__tests__/combatModifiers.test.ts
@@ -20,7 +20,7 @@ import {
   MANA_BLACK,
   MANA_WHITE,
 } from "@mage-knight/shared";
-import { addModifier } from "../modifiers.js";
+import { addModifier } from "../modifiers/index.js";
 import {
   DURATION_COMBAT,
   SCOPE_ONE_ENEMY,
