+++ packages/core/src/engine/__tests__/skillWhoNeedsMagic.test.ts
@@ -24,7 +24,7 @@ import { SKILL_TOVAK_WHO_NEEDS_MAGIC } from "../../data/skills/index.js";
 import { getValidActions } from "../validActions/index.js";
 import { getManaOptions } from "../validActions/mana.js";
 import { RULE_SOURCE_BLOCKED } from "../modifierConstants.js";
-import { isRuleActive } from "../modifiers.js";
+import { isRuleActive } from "../modifiers/index.js";
 import { sourceDieId } from "../../types/mana.js";
 import { MANA_RED, MANA_BLUE } from "@mage-knight/shared";
