+++ packages/core/src/engine/commands/endTurn/index.ts
@@ -17,7 +17,7 @@ import type { GameState } from "../../../state/GameState.js";
 import type { Player } from "../../../types/player.js";
 import type { GameEvent } from "@mage-knight/shared";
 import { TURN_ENDED, GAME_ENDED } from "@mage-knight/shared";
-import { expireModifiers } from "../../modifiers.js";
+import { expireModifiers } from "../../modifiers/index.js";
 import { EXPIRATION_TURN_END } from "../../modifierConstants.js";
 import { END_TURN_COMMAND } from "../commandTypes.js";
 import { createEndRoundCommand } from "../endRound/index.js";
